Updated openstack/openstack Project: openstack/tripleo-heat-templates  af1aed82693565da6a651c0b75a9f0d3d87a57b5
Containerize Redis for HA

This service allows configuring and deploying Redis containers
in a HA overcloud managed by pacemaker.

The containers are managed and run by pacemaker. Inside there is
pacemaker_remote which will invoke the resource agent managing galera.
The resources themselves are created via puppet-pacemaker inside a
short-lived container used for this purpose (mysql_init_bundle).
This container needs to use the 'docker_config' section to invoke
puppet (as opposed to 'docker_puppet_tasks'), because due to the HA
composability each resource creation needs to happen on the bootstrap
node of that service and 'docker_puppet_tasks' will only run on the
controller/primary role.

Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Closes-Bug: #1692924

Depends-On: Ia1131611d15670190b7b6654f72e6290bf7f8b9e

Change-Id: Ie045954fcc86ef2b3e4562b6f012853177f03948