Updated openstack/openstack Project: openstack-infra/tripleo-ci  e1b35eeb80a8ef046fef16951906933b91d5e164
Remove all centos-release-* repos

We shouldn't be using packages from any of these, and they can cause
package conflicts with the delorean repos we do want to use.

For some reason centos-release-* also matches the base centos-release
package, so we need to explicitly match only packages where the
following characters are a-z.

Change-Id: I928a0efc9d75e584889967e262155e576452632d
Closes-Bug: 1685814