Updated openstack/openstack Project: openstack/governance  21bf4a28a2e7dce5c07655e68b96dd2af51f3c70
Add etherpad link with detailed steps  to split-tempest-plugins goal

This commit adds a link to the etherpad which contains a detailed step
by step on how to create the repository and retain git history. These
steps are not a requirement to complete the goal, but can be used to
accomplish it.

Change-Id: I0b26cccd44ead20dca3ba05cf7b1b393211de797