Updated openstack/openstack Project: openstack-infra/project-config  b62ccc33e6f3a630925b7ec209792f8aac9bc18e
Test keystone with newly split tempest plugin

This patch adds a new nonvoting job for keystone that enables the
recently split out tempest plugin as per the guidelines[1] instead of
running the in-tree plugin. This job can become voting and replace the
main keystone-dsvm-functional job once we know this was set up
correctly.

[1] https://etherpad.openstack.org/p/tempest-separate-plugin

Change-Id: I0f37cadb88966ea572c3aba2265b195b0a6725fe