Updated openstack/openstack Project: openstack-infra/shade  4bb33a582f7dc16ee626d29ae8acad6ae1bf644f
Do less work when deleting a server and floating ips

When we delete a server with floating ips, we tell get_server to not
fetch a bare server, which does the work to fill in the network info
from neutron. Then we look in the server for the floating ip address and
look up the port it goes with.

This is not necessary.

We can tell get_server to get us a bare server, then look up floating
ips by device_id. Then just delete them.

Change-Id: I5ec04dc2a356aa20cf561866e8f43f9e28b2db21