Updated openstack/openstack Project: openstack/oslo.config  297ba5a1f04a7f9452e2c8eebb4fd1753a9e27ba
Cast generator_options to dict

At some point in the development of this feature, I cleaned up an
"unnecessary" loop that was writing the generator_options.  It turns
out this wasn't as unnecessary as I thought.  A ConfigOpts instance
looks dict-like enough to pass the unit tests, but when it's
actually passed to the yaml formatter it fails because yaml doesn't
know what to do with it.

Instead of reintroducing the loop, it seems just casting the
ConfigOpts to a dict will work.

Change-Id: I0aaac8485259cc80a430bb80684d1bfe1c7f91c3