Updated openstack/openstack Project: openstack/keystone  5258c340d695bb344d386f94900992a1e982dfce
Update DirectMappingError in keystone.exception

When the correct IndexError happens in federation mapping rule, '{0}'
is in a local section, but the value in direct_maps is null.

Now I have set insecure_debug to 'True', and it shows like '...
(e.g. {0} in a local section).'. But it still shows 'An unexpected...'.

Change-Id: If6263229b153828ffa07ee3ad6004f3db7cdfd98
Closes-Bug: #1695131