Updated openstack/openstack Project: openstack/tempest  a1d437276af80dbdd753c37607183e5ca379a697
Use assert(Not)Empty,IsNotNone instead of assert(Not)Equal,GreaterEqual(0, len(..

This commit makes to use assert(Not)Empty() or assertIsNotNone() instead
of assert(Not)Equal(0, len(LIST),..) or assertGreaterEqual(0, len()...).
The original code asserts the specified list is empty or not. And we
already have assert(Not)Empty() for it. The method names
"assert(Not)Empty" are straightforward and easy to understand. It makes
the code clean and the assert message could be easier to understand what
is happen than original ones.

Change-Id: I715ec84d9139b1d196275e91212ab40d52a545e1