Updated openstack/openstack Project: openstack/nova  820ad7d2a551b16f318de3f2f9d9e83c5642f120
Ensure image conversion flushes output data to disk

qemu-img convert defaults to cache=none, which means that when it
completes the output data may still only be in the host kernel's
cache rather than on persistent storage. A host crash at this point
will leave a file with the correct metadata (name, size, ownership,
permissions), but no contents. This will prevent qcow2 instances on
that compute host which use that image from restarting, and requires
manual intervention from an operator to fix.

See also change Id9905a87, which fixes this issue for downloads
without a conversion.

Closes-Bug: #1669844
Change-Id: I33bd99b0752111ff7057f9bd40e58dcde77c7d95