Updated openstack/openstack Project: openstack/nova  52189a50cd3342a3b3edd395a42e5a9a2cd5e768
Pass a list of instance UUIDs to scheduler

When the RequestSpec object was created, it was assumed that when the
request was for more than one instance the scheduler would not need to
know the UUIDs for the individual instances and so it was agreed to
only pass one instance UUID. If, however, we want the scheduler to be
able to claim resources on the selected host, we will need to know the
instance UUID, which will be the consumer_id of the allocation.

This patch adds a new RPC parameter 'instance_uuids' that will be passed
to the scheduler. The next patch in the series adds the logic to the
scheduler to use this new field when selecting hosts.

Co-Authored-By: Sylvain Bauza <sbauza@redhat.com>
Partially-Implements: blueprint placement-claims

Change-Id: I44ebdb3e29db950bf2ad0e6b1dbfdecd1ca03530