Updated openstack/openstack Project: openstack/tripleo-heat-templates  d18e1a6b64f63c74764bf88c44afa4d24e17775d
Containerize HAProxy for the non-ha case

This change implements an initial container for haproxy in the non-HA
case (aka when the container is not spawn by pacemaker).

We tested this using a stock kolla haproxy container image and we were
able to get haproxy running on a container with net=host correctly.

Change-Id: I90253412a5e2cd8e56e74cce3548064c06d022b1
Co-Authored-By: Michele Baldessari <michele@acksyn.org>
Depends-on: I51c482b70731f15fee4025bbce14e46a49a49938
Closes-Bug: #1668936