Updated openstack/openstack Project: openstack/tripleo-heat-templates  e797c0321b7758dd9a5554af2ea275c47df2a448
Container-specific variants of scenarios 001-004

Add container-specific variants of scenarios. These variants are
supposed to be temporary, as their only purpose is to allow us CIing the
scenarios with containers while we don't have pacemaker containerized
yet. Once we can deploy and upgrade containerized deployment with
pacemaker, these should be deleted and normal scenarios should be used.

Alternative approach would be to edit the scenarios on the fly within
the CI job, to remove the pacemaker parts, which would be more DRY, but
perhaps more surprising when trying to debug issues.

Change-Id: I36ef3f4edf83ed06a75bc82940152e60f9a0941f