Updated openstack/openstack Project: openstack-infra/devstack-gate  bccf691bd1ce4597260c09ab2ad9873ae2f2f458
List swift and keystone in feature.yaml

This lists swift and keystone as configurations in the features.yaml
file so that they can be toggled on using DEVSTACK_GATE_SWIFT and
DEVSTACK_GATE_KEYSTONE variables. When paired with
DEVSTACK_GATE_NO_SERVICES this allows us to easily only turn on keystone
and swift for swift functional testing.

This isn't strictly necessary but makes the swift functional test jobs
run much quicker.

Change-Id: I1ace341721487e4700de308d3d5b800661210fc0