Updated openstack/openstack Project: openstack-infra/project-config  04ad13dd1d79cf1d94e45714e9218ccdd0188973
Only run swift and keystone in swift func jobs

Previously the swift functional jobs ran a default devstack install.
This results in a cloud that works for running the swift test suite, but
require additional setup time that is unnecessary. Only install keystone
and swift as the other services aren't required in order to save job
runtime.

Depends-On: I1ace341721487e4700de308d3d5b800661210fc0
Change-Id: If169ac7c1b595c69e5800a3c88edd60547b577f0