Updated openstack/openstack Project: openstack/nova  11ad2da58b7746627faff7083689706004826b9d
Target cell on local delete

This fixes an issue where we don't actually target the cell if we're
doing a local delete. We might have the cell from earlier in the checks,
but if not, we look it up and use it.

We can't hit this if we have a single cell, but we do if we're running
on multi-cell and do the "delete while booting" tests in tempest. Indeed,
this issue was discovered whilst working through the backlog of failures
with the dependent devstack change to enable multi-cell.

Change-Id: I718ca99fcef7e83fe1955685ca4c3366df02efe8