Updated openstack/openstack Project: openstack/cinder  1eccae469ce0e1aec6063947deb7312b7315d08c
Mark Reduxio driver as unsupported

The Reduxio CI has not reported success past our CI required period.

Checking name: Reduxio HX550 CI
  first seen: 2017-03-14 02:25:08 (51 days, 14:46:16 old)
  last seen: 2017-03-23 06:34:03 (42 days, 10:37:21 old)
  last success:  None
  Job dsvm-cinder-driver-copy 0% success out of 250 comments S=0, F=250
Overall success rate: 0% of 250 comments

Per Cinder's non-compliance policy, this patch marks the driver
as unsupported and deprecated and it will be removed in the
next cycle if the issue is not corrected.

Change-Id: I6a01344be588d507eba739bd1a6e6a47ea3f159c