Updated openstack/openstack Project: openstack-infra/project-config  b66c410ad3fbda470350a9d0c9d3c265ca7d2d32
do not use the remote prefix name for release notes scans

Leave the remote name out of the branch specifier for new release notes
pages to allow local dev/test builds to work. Reno will search for a
remote branch if it does not find a local branch with the name given.

Change-Id: If85133edcc45d6fc9c7fcca6fc45c647ba5f20a1
Signed-off-by: Doug Hellmann <doug@doughellmann.com>