Updated openstack/openstack Project: openstack-infra/project-config  a7607530b37da4db9d72aa7ccd5a366ded8699e9
jenkins: propose_update.sh: Use a single active repository for testing

Previously, we used a retired repo to verify that the job will not do
any updates on repositories which have no matching files. We now change
that to a single active repository to confirm that the job is working
properly under normal conditions.

Change-Id: I0617090370eed20676239490c0846b4e28ce6812