Updated openstack/openstack Project: openstack/ironic-python-agent  da2ee32a8b2c15a232397e4ed5bc01b695cba00c
[imagebuild] support requirements referring to /opt/stack/new

Now that ironic-lib builds IPA from source, the IPA building procedure
needs to support requirements referring to /opt/stack/new.

As upper-constraints are not actually needed in the finalise step, they're
no longer used there to avoid dealing with file:/// references. We also
allow pre-release packages to be installed by pip.

Also install git as it's a build requirement for PBR.

Only tinyipa is updated, as it's used in the ironic-lib gate.

Change-Id: Ia942cb710eb3fe1ba56c68f91fa893795c9cb651
Closes-Bug: #1695877