Updated openstack/openstack Project: openstack/nova  8d5c1608f8ea08272119ce46242e97c58309b1f0
Fix cell0 naming when QS params on the connection

We had a problem when the nova connection string included parameters on the
query string like charset encoding.
Note that the connection string necessarly needs to be RFC1738 compliant as
per Sqlalchemy rules, so it's totally safe to just unquote what the SQLA
helper method gives us as a result.

Also removed a tested connection string since it wasn't RFC1738 compatible.

Change-Id: I45fe9b92e8d93a0099d33bb0070e9d4e540595ac
Closes-Bug: #1696001