Updated openstack/openstack Project: openstack/cinder  3e0b7c362b955cd58daf0ddf48b05a7d57fc148e
Create volumes from encrypted images

Volume encryption keys specified in Glance image
metadata are copied to new keys for volumes cloned
from that image.

If the image key and volume key match (i.e. when using
the conf key manager), the data is copied directly from
the image to the volume.

When creating a new volume in a Barbican environment, we
will generate a new encryption key which is identical to
the original key.

If creating an unencrypted volume, the encrypted image data
is copied into the volume.  This is not directly usable
without the encryption key, but may be useful for data transfer
purposes.

Closes-Bug: #1485449
Related bp: improve-encrypted-volume

Change-Id: I1f4ea35cd05f4c43a5ae07d8a541ff6495d5f8e9