Updated openstack/openstack Project: openstack/horizon  c9f073930b2ea1ad5adc84f980fd7d5b45243a26
Fix instance launch dialog under MSIE11

The novaAPI.getFlavors() call takes only a single parameter, which
should be an object with params. But it is being called with two
parameters here, both "true", which makes IE11 fail with TypeError
later on, when it tries to iterate over keys of params.

Change-Id: I0dcfcdca3dda2d94509d7a574184e7f82f21ca22
Closes-Bug: #1694574