Updated openstack/openstack Project: openstack/tripleo-heat-templates  671db66b9c18a809dbdbabbf48823754a2402bcf
Convert puppet and docker steps to ansible

Replace the multiple SoftwareDeployment resources with a common
playbook that runs on all roles, consuming the configuration data
written via the HostPrepAnsible tasks.

This hopefully simplifies things, and will enable re-running the
deploy steps for minor updates (we'll need some way to detect
a container should be replaced, but that will be done via a
follow-up patch).

Change-Id: I674a4d9d2c77d1f6fbdb0996f6c9321848e32662