Updated openstack/openstack Project: openstack/cinder  ba9711d17201223999e0879930efe8523de6f405
Mark Nimble driver as unsupported

The Nimble CI has not reported success past our CI required
period.

Checking name: Nimble Storage CI
  first seen: 2016-07-12 18:33:36 (295 days, 22:36:45 old)
  last seen: 2017-02-27 23:42:16 (65 days, 17:28:05 old)
  last success: 2017-02-27 23:42:16 (65 days, 17:28:05 old)
  Job nimble-fc-driver-dsvm-volume 61% success out of 88 comments S=54, F=34
  Job nimble-iscsi-driver-dsvm-volume 36% success out of 131 comments S=48, F=83
      last success: 2017-01-26 20:33:12 (97 days, 20:37:09 old)
Overall success rate: 46% of 250 comments

Per Cinder's non-compliance policy, this patch marks the driver
as unsupported and deprecated and it will be removed in the
next cycle if the issue is not corrected.

Change-Id: Ie068ae8d8fbc224d78118818426e514ba3e100e3