Updated openstack/openstack Project: openstack/keystone  c528539879e824b8e6d5654292a85ccbee6dcf89
Quotation marks should be included in http url using curl

I think using qutation marks in http url is more regular, even though
the current examples in the doc have no problem without qutation marks.
The doc is a manual of api curl, a user is likely to refer to
the manual and curl an another api which not in the doc.

For example:
curl -s \
     -H "X-Auth-Token:$OS_TOKEN" \
     http://loaclhost:5000/v3/role_assignments?user.id=$USER_ID&scope.project.id=$PROJECT_ID
response:
[1] 22894
-bash: scope.project.id=4f91eb36955543c9826b705bc5a90c59:command not found
...
The url has special characters such as '&', which can not be handled correctly.
Using double quotation marks in http url will be ok, like this:
"http://loaclhost:5000/v3/role_assignments?user.id&scope.project.id"

Change-Id: Id8f4c3cc64d298975ad81f8fc08abbec920a1e09