Updated openstack/openstack Project: openstack/nova  c7afbb95e67f5392845c24cf3d69345759aafde2
Get BDMs when we need to in _handle_cell_delete

The BDM list was originally added in this flow
in change Id68330eb238a304eaceefa7c0ec2fa0d2b60d28a
and it's only used for the _local_delete call but
over time with the cells v2 code it's been moved
away. This change moves the BDM query back to
right before _local_delete so we don't make an
unnecessary call in a semi-time sensitive code
path (because of the race between deleting the
instance in the API and the conductor deleting
the build request).

Change-Id: I6f4055adbc33701a39cf2b92eff101083a4f56f6