Updated openstack/openstack Project: openstack/nova  6c58884030a016bc88ee0019902a64937b1dbced
Remove **kwargs passing in payload __init__

The **kwargs was used in the payload class ctor to pass the payload
fields initialization up to the ovo base class. This is unnecessary
and confusing.

This patch replaces the **kwargs passing with direct field initialization
to make the payload construction more explicit.

Change-Id: I7770f6550fa40c1d9b0417efa57f58fa9c998d45
Implements: bp versioned-notification-transformation-pike
Co-Authored-By: Stephen Finucane <sfinucan@redhat.com>