Updated openstack/openstack Project: openstack/docs-specs  f0fe465f695ff7b4c1a246a07eae877746928e46
Adds spec for migrating the documentation suite to project repos

Currently the user-facing documentation belongs in the
openstack-manuals repo. Due to a severe decline in contributions,
the docs team has had to come up with a new way to move forward.

This spec proposes the new process for user-facing documentation
and how we are going to undertake the changes.

Co-Authored By: Doug Hellmann <doug@doughellmann.com>
Co-Authored By: Alexandra Settle <a.settle@outlook.com>
Change-Id: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454