Updated openstack/openstack Project: openstack/nova  480f5c1c4e857b11b317732814428de09441b445
Default to 0 when merging values in limit check

When merging project_values and user_values after taking the symmetric
difference of their keys, we should use a default value of 0 for the
.get(key) calls because a default value of None will not take
precedence over 0 if 0 is the value in project_values or user_values.
In the buggy case, we end up with a merged value of None instead of 0
because: 'None or 0' is 0 but '0 or None' is None.

Closes-Bug: #1696605

Change-Id: I54a3a52cf5e9b7826dad37c6d1e053ba92cd4fd9