Updated openstack/openstack Project: openstack/horizon  ad21f99dfa8fa1a89b7c17fe2125d86565c3cbac
Fixing types.MethodType signature for Python 3.5

The signature for MethodType has changed in Python 3,
and the current implementation won't work, to fix this
we are replacing it with the py 2/3 compatible
python-six equivalent, six.create_unbound_method.
http://six.readthedocs.io/#six.create_unbound_method

This is causing the Python 3.5 tests to fail for the
octavia-dashboard when importing the navigation.py
file, throwing the following error.
> TypeError: method expected 2 arguments, got 3

The following review to fix the gate for the
octavia-dashboard is depending on this patch.
https://review.openstack.org/#/c/471611/

This can also be reproduced by simply executing
navigation.py using Python 3.5.
> python3.5 nagivation.py

Change-Id: Iebca15ae962a1aa0825c6ba6d482206531726696