Updated openstack/openstack Project: openstack/nova  ede7fd48dcce8a2dca6a9920be8bee6a7696a9fe
Use 3.27 microversion when creating new style volume attachments

Now that we have version discovery support we can explicitly
request 3.27 when creating a new-style volume attachment. If the
3.27 microversion is not available, CinderAPIVersionNotAvailable
will be raised back up to the caller, indicating an old-style
attachment needs to be made.

The same is done when updating or deleting new-style attachments.

A TODO is left inline for update and delete since we should be able
to assume that if we're updating or deleting a new style attachment,
which is keyed off the BlockDeviceMapping.attachment_id field, that
3.27 is available and we can save ourselves a GET / call to Cinder.

Part of blueprint cinder-new-attach-apis

Change-Id: Icf8cda2c34dda9e7b56a586353842dd035b115cd