Updated openstack/openstack Project: openstack/neutron  b6bd47562985528bc8bbc0586c1da2e08cf64f3f
api-tests: Common way to define required extensions

As some tests require extensions to be enabled on the server, this patch
introduces a new class attribute to the API base class. If any extension
defined in test class is not enabled on the server, then all tests in
the test class will be skipped.

Change-Id: I0629153f973daeb3bef3a6968360cbc2d427f9ad