Updated openstack/openstack Project: openstack/nova  2c69304c150568dccb2292a49c1fd4bb9de9d3ef
Add get_minimum_version_all_cells() helper for service

If we're making decisions in the API about the age of services
such as compute, we need to survey all the cells to get the minimum
across all of them. This adds a helper function that can be used
from places where we need to do that.

Change-Id: I9976de405de575d9da1389e629c17694207d8c34
Related-Bug: #1695993