Updated openstack/openstack Project: openstack/nova  c07cca7f9153a7943fcffd13bac1cdf06029a9cf
Add InstanceGroup._remove_members_in_db

This adds a method for removing server group members from the database.
It's private because we don't expose a way to remove server group
members in the API and it's only needed to handle races near the quota
limit. To handle races, we will check quota limits, create group
members in the database, then check quota again. If the quota limits
have been exceeded after creation of the group members, they will
be removed from the database and OverQuota will be raised.

Part of blueprint cells-count-resources-to-check-quota-in-api

Change-Id: I8cda7a1503750f6b23919a6fbf7bb40c84ac582b