Updated openstack/openstack Project: openstack/nova  10966e962e1023d379915185e7ed433a3d748258
Prefer non-PCI host nodes for non-PCI instances

When trying to fit an instance's topology to a host's, various
permutations of the host are attempted in the hope of finding one that
can be used. Modify this iteration to try use cells without a PCI
device attached first when we don't require PCI devices.

Change-Id: I8be6bf8c437fc801499b213c2e884defb7900b17
Implements: blueprint reserve-numa-with-pci