Updated openstack/openstack Project: openstack-infra/system-config  aec4d5fb3980f7263ba0b4c52236f9586309e5ba
Run vos release only when necessary

Through testing we have discovered that every vos release affects remote
AFS caches. It appears that doing a vos release (even one that is a nop)
causes remote AFS clients to update file metadata info on file read.
This affects are pypi mirror in particular because we vos release on it
every 5 minutes.

Make this less costly in the pypi mirror by checking which packages are
updated by bandersnatch against our upper constraints list and only
vos release if the package is in upper constraints. We will also vos
release if 4 hours has passed since the last vos release to catch any
packages not listed in constraints.

Change-Id: Ib79393300111ba300cbea92200b6d8e729672354